#11: BackendConfigUi moved to ui subpackage
