+++ src/main/scala/org/fs/mael/core/backend/Backend.scala
@@ -4,6 +4,7 @@ import java.io.File
 import java.net.URI
 
 import org.eclipse.swt.widgets.TabFolder
+import org.fs.mael.core.backend.ui.BackendConfigUi
 import org.fs.mael.core.checksum.Checksum
 import org.fs.mael.core.config.InMemoryConfigStore
 import org.fs.mael.core.entry.DownloadEntry
