+++ src/main/scala/org/fs/mael/core/backend/ui/AbstractBackendConfigUi.scala
@@ -1,4 +1,4 @@
-package org.fs.mael.core.backend
+package org.fs.mael.core.backend.ui
 
 import org.eclipse.swt.widgets.TabFolder
 import org.fs.mael.core.config.ConfigStore
